DRILL-6656: Disallow extra semicolons and multiple statements on the same line.
closes #1415

+++ logical/src/main/java/org/apache/drill/common/expression/BooleanOperator.java
@@ -59,7 +59,7 @@ public class BooleanOperator extends FunctionCall{
       i++;
     }
 
-    return (int) (cost / i) ;
+    return cost / i;
   }
 
 }
